Remove cast to double in things like a = b / c where a, b, and c are all
2005-11-23  Behdad Esfahbod  <behdad@gnome.org>

        * pango/mapping.c (pango_glyph_string_index_to_x),
        pango/pango-layout.c (pango_layout_iter_get_char_extents):
        Remove cast to double in things like a = b / c where a, b,
        and c are all ints. (#312469, Sven Neumann)

+++ pango/mapping.c
@@ -148,9 +148,8 @@ pango_glyph_string_index_to_x (PangoGlyphString *glyphs,
   if (trailing)
     cluster_offset += 1;
 
-  *x_pos = (((double)(cluster_chars - cluster_offset)) * start_xpos +
-	    ((double)cluster_offset) * end_xpos) /
-    cluster_chars;
+  *x_pos = ((cluster_chars - cluster_offset) * start_xpos +
+	    cluster_offset * end_xpos) / cluster_chars;
 }
 
 /**
